requirements.txt
tabula-py
psycopg2==2.9.7